实战项目 - 论坛 - 23、实现评论功能（不刷新页面一二级评论，点赞）

+++ 3、实战项目/community/src/main/java/life/majiang/community/dto/QuestionDTO.java
@@ -1,19 +1,24 @@
 package life.majiang.community.dto;
 
+import life.majiang.community.model.Comment;
 import life.majiang.community.model.User;
 import lombok.Data;
 
+import java.util.List;
+
 @Data
 public class QuestionDTO {
-    private Integer id;
+
+    private Long id;
     private String title;
     private String description;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;
+    private Long creator;
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
     private String tag;
     private User user;
+
 }
